refactor(editor): simplify CollaborativeEditor and enhance editor upload handling
- Extract WebSocket provider logic to useWebsocketProvider hook for cleaner state management
- Remove deprecated debounce function, replace auto-save with useAutoSave hook
- Move initial content loading logic to useInitialContent hook for better separation of concerns
- Add file upload support via uploadFiles function to CollaborativeEditor and Editor components
- Introduce debounced onChange callback in Editor for optimized update handling
- Replace manual WebSocket reconnection with managed provider state and events
- Improve CollaborativeEditor code readability by removing unused imports and states
- Fix loading state display in IdeaInfo component during data fetching
