+++ apps/web/components/collaborative-editor.tsx
@@ -1,101 +1,19 @@
 "use client";
 
-import { BlockNoteEditor, PartialBlock } from "@blocknote/core";
+import { BlockNoteEditor } from "@blocknote/core";
 import { BlockNoteView } from "@blocknote/mantine";
 import { useCreateBlockNote } from "@blocknote/react";
 import "@blocknote/core/fonts/inter.css";
 import "@blocknote/mantine/style.css";
-import { useEffect, useState, useCallback, useRef, useMemo } from "react";
+import { useEffect, useState, useRef } from "react";
 import * as Y from "yjs";
-import { WebsocketProvider } from "y-websocket";
-import {
-  saveEntityDocumentContent,
-  getEntityDocumentContent,
-} from "../actions/documents/document";
 import { useTheme } from "next-themes";
 import { useSession } from "../context/session-context";
-import { useQuery } from "@tanstack/react-query";
-import { queryKeys } from "../lib/query-keys";
-
-type DocumentEntityType =
-  | "project"
-  | "issue"
-  | "feature"
-  | "milestone"
-  | "competitor"
-  | "competitorSwot"
-  | "competitiveMove"
-  | "roadmapItem"
-  | "actionItem";
-
-// Custom debounce function
-function debounce<T extends (...args: any[]) => any>(
-  func: T,
-  wait: number
-): T & { cancel: () => void } {
-  let timeout: NodeJS.Timeout | null = null;
-
-  const debounced = ((...args: Parameters<T>) => {
-    if (timeout) clearTimeout(timeout);
-    timeout = setTimeout(() => func(...args), wait);
-  }) as T & { cancel: () => void };
-
-  debounced.cancel = () => {
-    if (timeout) {
-      clearTimeout(timeout);
-      timeout = null;
-    }
-  };
-
-  return debounced;
-}
-
-interface CollaborativeEditorProps {
-  // New API: save by entity type and id
-  entityType?: DocumentEntityType;
-  entityId?: string;
-}
-
-const USER_COLORS = [
-  "#FF6B6B",
-  "#4ECDC4",
-  "#45B7D1",
-  "#96CEB4",
-  "#FFEAA7",
-  "#DDA0DD",
-  "#98D8C8",
-  "#F7DC6F",
-  "#BB8FCE",
-  "#85C1E9",
-  "#FF9F43",
-  "#54A0FF",
-  "#5F27CD",
-  "#00D2D3",
-  "#FF9FF3",
-  "#54A0FF",
-  "#1DD1A1",
-  "#FD79A8",
-  "#FDCB6E",
-  "#6C5CE7",
-  "#A29BFE",
-  "#FD79A8",
-  "#E17055",
-  "#00B894",
-  "#00CEC9",
-  "#E84393",
-  "#FDCB6E",
-  "#6C5CE7",
-  "#74B9FF",
-  "#55A3FF",
-];
-
-function getUserColor(userId: string): string {
-  let hash = 0;
-  for (let i = 0; i < userId.length; i++) {
-    hash = userId.charCodeAt(i) + ((hash << 5) - hash);
-  }
-  return USER_COLORS[Math.abs(hash) % USER_COLORS.length] || "#000000";
-}
+import { CollaborativeEditorProps, getUserColor } from "./editor/index";
+import { useWebsocketProvider } from "./editor/use-websocket-provider";
+import { useAutoSave } from "./editor/use-auto-save";
+import { useInitialContent } from "./editor/use-initial-content";
+import { uploadFiles } from "@/lib/uploadthing";
 
 export function CollaborativeEditor({
   entityType,
@@ -103,160 +21,31 @@ export function CollaborativeEditor({
 }: CollaborativeEditorProps) {
   const { theme } = useTheme();
   const [doc] = useState(() => new Y.Doc());
-  const [provider, setProvider] = useState<WebsocketProvider | null>(null);
-  const [connectionStatus, setConnectionStatus] = useState<
-    "connecting" | "connected" | "disconnected"
-  >("connecting");
   const editorRef = useRef<BlockNoteEditor | null>(null);
-  const reconnectTimeoutRef = useRef<NodeJS.Timeout | null>(null);
   const user = useSession();
 
-  // Create stable room ID (org scoping disabled for dev mode)
-  const stableRoomId = useMemo(() => {
-    const baseRoomId = `${entityType}-${entityId}` || "default-room";
-    // Organization scoping disabled for development
-    return baseRoomId;
-  }, [entityType, entityId]);
+  const { provider, connectionStatus, setConnectionStatus } =
+    useWebsocketProvider(entityType, entityId, doc);
 
-  // Initialize provider once and handle cleanup
-  useEffect(() => {
-    // WebSocket initialization (authentication disabled for dev mode)
+  const debouncedSave = useAutoSave(entityType, entityId);
 
-    let websocketProvider: WebsocketProvider | null = null;
-    let mounted = true;
+  const uploadFile = async (file: File) => {
+    try {
+      const res = await uploadFiles("fileUpload", { files: [file] });
+      const uploaded = Array.isArray(res) ? res[0] : res;
+      return uploaded?.ufsUrl || "";
+    } catch (error) {
+      console.error("Error uploading file:", error);
+      return "";
+    }
+  };
 
-    const initProvider = async () => {
-      try {
-        // Build WebSocket URL with room path (dev mode)
-        const baseUrl =
-          process.env.NEXT_PUBLIC_WEBSOCKET_URL || "ws://localhost:1234";
-        const wsUrl = `${baseUrl}/${stableRoomId}`;
-
-        websocketProvider = new WebsocketProvider(wsUrl, stableRoomId, doc);
-
-        // Add connection event listeners for debugging
-        websocketProvider.on("status", (event: any) => {
-          console.log("WebSocket status:", event.status);
-          if (mounted) {
-            if (event.status === "connected") {
-              setConnectionStatus("connected");
-            } else if (event.status === "disconnected") {
-              setConnectionStatus("disconnected");
-            }
-          }
-        });
-
-        websocketProvider.on("connection-close", (event: any) => {
-          console.log("WebSocket connection closed:", event);
-          if (mounted) {
-            setConnectionStatus("disconnected");
-            // Auto-reconnect after a delay
-            reconnectTimeoutRef.current = setTimeout(() => {
-              if (mounted && websocketProvider) {
-                console.log("Attempting to reconnect WebSocket...");
-                setConnectionStatus("connecting");
-                websocketProvider.connect();
-              }
-            }, 3000); // Reconnect after 3 seconds
-          }
-        });
-
-        websocketProvider.on("connection-error", (event: any) => {
-          console.error("WebSocket connection error:", event);
-          if (mounted) {
-            setConnectionStatus("disconnected");
-            // Auto-reconnect after a delay on error
-            reconnectTimeoutRef.current = setTimeout(() => {
-              if (mounted && websocketProvider) {
-                console.log("Attempting to reconnect after error...");
-                setConnectionStatus("connecting");
-                websocketProvider.connect();
-              }
-            }, 5000); // Longer delay for errors
-          }
-        });
-
-        // Small delay to allow initial connection and prevent rapid reconnection cycles
-        await new Promise((resolve) => setTimeout(resolve, 500));
-
-        if (mounted) {
-          setProvider(websocketProvider);
-        }
-      } catch (e) {
-        console.error("Failed to create WebsocketProvider:", e);
-        if (mounted) {
-          setProvider(null);
-          setConnectionStatus("disconnected");
-        }
-      }
-    };
-
-    initProvider();
-
-    return () => {
-      mounted = false;
-      if (reconnectTimeoutRef.current) {
-        clearTimeout(reconnectTimeoutRef.current);
-      }
-      if (websocketProvider) {
-        websocketProvider.destroy();
-      }
-    };
-  }, [stableRoomId, doc]);
-
-  // Auto-save functionality with debouncing
-  const debouncedSave = useCallback(
-    debounce(async (content: PartialBlock[]) => {
-      // Determine save strategy: entity-based preferred, fallback to legacy documentId
-      const resolved: {
-        type?: DocumentEntityType;
-        id?: string;
-      } = {
-        type: entityType,
-        id: entityId,
-      };
-
-      try {
-        let result:
-          | { success: true; document: any }
-          | { success: false; error: string };
-
-        if (resolved.type && resolved.id) {
-          result = await saveEntityDocumentContent({
-            entityType: resolved.type,
-            entityId: resolved.id,
-            content,
-          });
-        } else {
-          return;
-        }
-
-        if (result.success) {
-          console.log("Document saved successfully:", result.document.id);
-        } else {
-          console.error("Failed to save document:", result.error);
-        }
-      } catch (error) {
-        console.error("Auto-save failed:", error);
-      } finally {
-      }
-    }, 2000),
-    [entityType, entityId]
-  );
-
-  const { data: initialContent, isLoading: isFetchingInitialContent } =
-    useQuery({
-      queryKey: queryKeys.documents.entity(entityType, entityId),
-      queryFn: async () => {
-        if (!entityType || !entityId) return null;
-        const result = await getEntityDocumentContent({ entityType, entityId });
-        if (result.success) {
-          return result.document.content as PartialBlock[];
-        }
-        return null;
-      },
-      enabled: !!entityType && !!entityId,
-    });
+  const {
+    initialContent,
+    isFetchingInitialContent,
+    initialContentLoaded,
+    setInitialContentLoaded,
+  } = useInitialContent(entityType, entityId);
 
   const editor = useCreateBlockNote(
     {
@@ -272,7 +61,9 @@ export function CollaborativeEditor({
               showCursorLabels: "always",
             }
           : undefined,
+      uploadFile,
     },
+
     [provider, user]
   );
 
@@ -283,9 +74,6 @@ export function CollaborativeEditor({
       setTimeout(() => setConnectionStatus("connected"), 500);
     }
   }, [editor, provider, connectionStatus]);
-
-  // State to track if initial content has been loaded
-  const [initialContentLoaded, setInitialContentLoaded] = useState(false);
 
   // Effect to load initial content into the editor
   useEffect(() => {
@@ -335,7 +123,10 @@ export function CollaborativeEditor({
     editor.onChange(handleChange);
 
     return () => {
-      debouncedSave.cancel();
+      // Cancel any pending debounced save when component unmounts
+      if (typeof debouncedSave.cancel === "function") {
+        debouncedSave.cancel();
+      }
     };
   }, [editor, debouncedSave, connectionStatus]);
